Benchmarks: producer writing 500 bytes messages
