+++ benchmarks/src/it/scala/akka/kafka/benchmarks/NoCommitBackpressure.scala
@@ -15,21 +15,22 @@ class RawKafkaCommitEveryPollConsumer extends BenchmarksBase() {
                 KafkaConsumerBenchmarks.consumerAtLeastOnceCommitEveryPoll())
   }
 
-  it should "bench with normal messages" in {
-    val cmd = RunTestCommand(prefix + "consumer-normal-msg", bootstrapServers, topic_1000_5000)
-    runPerfTest(cmd,
-                KafkaConsumerFixtures.filledTopics(cmd),
-                KafkaConsumerBenchmarks.consumerAtLeastOnceCommitEveryPoll())
-  }
-
-  it should "bench with normal messages and eight partitions" in {
-    val cmd = RunTestCommand(prefix + "consumer-normal-msg-8-partitions",
-                             bootstrapServers,
-                             topic_1000_5000_8)
-    runPerfTest(cmd,
-                KafkaConsumerFixtures.filledTopics(cmd),
-                KafkaConsumerBenchmarks.consumerAtLeastOnceCommitEveryPoll())
-  }
+// These are not plotted anyway
+//  it should "bench with normal messages" in {
+//    val cmd = RunTestCommand(prefix + "consumer-normal-msg", bootstrapServers, topic_1000_5000)
+//    runPerfTest(cmd,
+//                KafkaConsumerFixtures.filledTopics(cmd),
+//                KafkaConsumerBenchmarks.consumerAtLeastOnceCommitEveryPoll())
+//  }
+//
+//  it should "bench with normal messages and eight partitions" in {
+//    val cmd = RunTestCommand(prefix + "consumer-normal-msg-8-partitions",
+//                             bootstrapServers,
+//                             topic_1000_5000_8)
+//    runPerfTest(cmd,
+//                KafkaConsumerFixtures.filledTopics(cmd),
+//                KafkaConsumerBenchmarks.consumerAtLeastOnceCommitEveryPoll())
+//  }
 }
 
 class AlpakkaCommitAndForgetConsumer extends BenchmarksBase() {
@@ -42,19 +43,20 @@ class AlpakkaCommitAndForgetConsumer extends BenchmarksBase() {
                 ReactiveKafkaConsumerBenchmarks.consumerCommitAndForget(commitBatchSize = 1000))
   }
 
-  it should "bench with normal messages" in {
-    val cmd = RunTestCommand(prefix + "normal-msg", bootstrapServers, topic_1000_5000)
-    runPerfTest(cmd,
-                ReactiveKafkaConsumerFixtures.committableSources(cmd),
-                ReactiveKafkaConsumerBenchmarks.consumerCommitAndForget(commitBatchSize = 1000))
-  }
-
-  it should "bench with normal messages and eight partitions" in {
-    val cmd = RunTestCommand(prefix + "normal-msg-8-partitions",
-                             bootstrapServers,
-                             topic_1000_5000_8)
-    runPerfTest(cmd,
-                ReactiveKafkaConsumerFixtures.committableSources(cmd),
-                ReactiveKafkaConsumerBenchmarks.consumerCommitAndForget(commitBatchSize = 1000))
-  }
+// These are not plotted anyway
+//  it should "bench with normal messages" in {
+//    val cmd = RunTestCommand(prefix + "normal-msg", bootstrapServers, topic_1000_5000)
+//    runPerfTest(cmd,
+//                ReactiveKafkaConsumerFixtures.committableSources(cmd),
+//                ReactiveKafkaConsumerBenchmarks.consumerCommitAndForget(commitBatchSize = 1000))
+//  }
+//
+//  it should "bench with normal messages and eight partitions" in {
+//    val cmd = RunTestCommand(prefix + "normal-msg-8-partitions",
+//                             bootstrapServers,
+//                             topic_1000_5000_8)
+//    runPerfTest(cmd,
+//                ReactiveKafkaConsumerFixtures.committableSources(cmd),
+//                ReactiveKafkaConsumerBenchmarks.consumerCommitAndForget(commitBatchSize = 1000))
+//  }
 }
